extract logic.ts from panels/filters.ts and relocate tests
Move pure helpers (applyFilters, matchesFilter, matchesDateFilter,
filterSummary, newFilterId, collectAllTags, nowForField,
makeInitialFilterState) into panels/filters/logic.ts. Relocate the
existing test file alongside it as panels/filters/filters.test.ts and
point its imports at ./logic.ts and ./types.ts. panels/filters.ts
re-exports the helpers so external call sites are unchanged.
